refactor: renaming package to envreader
