+++ envreader/type.go
@@ -1,0 +1,6 @@
+package envreader
+
+type EnvReader struct {
+	Errs []string
+	Env  func(key string) string
+}
