--- readenv/type.go
@@ -1,6 +1,0 @@
-package readenv
-
-type EnvReader struct {
-	Errs []string
-	Env func(key string) string
-}
